renamed autoboost app
