--- apps/accelerate/autoboost/autoboost/__init__.py
@@ -1,1 +1,0 @@
-from autoboost.api.functions import optimize_model  # noqa: F401
