+++ apps/accelerate/speedster/requirements.txt
