+++ apps/accelerate/speedster/speedster/__init__.py
@@ -1,0 +1,1 @@
+from speedster.api.functions import optimize_model  # noqa: F401
